# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Huginn API
  version: 0.1.0
  description: |
    Initial contract scaffold for Huginn's HTTP API.
    This spec currently focuses on core session/chat, agent, workflow, connection,
    and system endpoints used by the web app. It is intended to be expanded
    incrementally as additional routes are hardened.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: List installed skills
      responses:
        "200":
          description: Installed skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillItem"
    post:
      tags: [Skills]
      summary: Create skill from markdown content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "201":
          description: Skill created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                required: [name]
        "400":
          $ref: "#/components/responses/ErrorResponse"
...
  /api/v1/skills/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema:
          type: string
    get:
      tags: [Skills]
      summary: Get installed skill details
      responses:
        "200":
          $ref: "#/components/responses/GenericObjectResponse"
    put:
      tags: [Skills]
      summary: Update skill markdown
      responses:
        "200":
          $ref: "#/components/responses/GenericObjectResponse"
    delete:
      tags: [Skills]
      summary: Uninstall skill
      responses:
        "204":
          description: Skill removed
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HuginnToken
  responses:
    ErrorResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    GenericObjectResponse:
      description: Generic success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenericObject"
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SkillItem:
      type: object
      properties:
        name:
          type: string
        author:
          type: string
        source:
          type: string
        enabled:
          type: boolean
        tool_count:
          type: integer
      required: [name, enabled, tool_count]